+++ test/lib/puma_ctl_spec.rb
@@ -43,6 +43,55 @@ describe PumaCtl do
     end
   end
 
+  describe "#gc_stats" do
+    let(:control_auth_token) {'123'}
+    let(:state_file_path) {nil}
+    let(:socket) {spy('socket')}
+
+    before do
+      allow(socket).to receive(:read).and_return(%Q( HTTP\/1.1 200 OK\r\n{"heap_used": 1516}))
+      allow(Socket).to receive(:tcp).and_yield(socket)
+      allow(Socket).to receive(:unix).and_yield(socket)
+    end
+
+    context 'with a TCP control url' do
+      let(:control_url) {'tcp://192.1.1.25:9191'}
+
+      it 'connects to the TCP socket' do
+        puma_ctl.gc_stats
+        expect(Socket).to have_received(:tcp).with('192.1.1.25', '9191')
+      end
+
+      it 'sends the correct command to the control server' do
+        expect(socket).to receive(:print).with("GET /gc-stats?token=#{control_auth_token} HTTP/1.0\r\n\r\n")
+        puma_ctl.gc_stats
+      end
+
+      it 'returns the JSON socket response as a hash' do
+        expect(puma_ctl.gc_stats).to eq("heap_used" => 1516)
+      end
+    end
+
+    context 'with a UNIX socket control url' do
+      let(:control_url) {'unix:///some/path'}
+
+      it 'connects to the socket' do
+        puma_ctl.gc_stats
+        expect(Socket).to have_received(:unix).with('/some/path')
+      end
+
+      it 'sends the correct command to the control server' do
+        expect(socket).to receive(:print).with("GET /gc-stats?token=#{control_auth_token} HTTP/1.0\r\n\r\n")
+        puma_ctl.gc_stats
+      end
+
+      it 'returns the JSON socket response as a hash' do
+        expect(puma_ctl.gc_stats).to eq("heap_used" => 1516)
+      end
+    end
+
+  end
+
   describe "#stats" do
     let(:control_auth_token) {'123'}
     let(:state_file_path) {nil}